use the real RTCPeer implementation
