+++ packages/swaig/package.json
@@ -37,10 +37,11 @@
     "prepublishOnly": "npm run build"
   },
   "dependencies": {
-    "@fastify/basic-auth": "^6.0.1",
-    "@fastify/swagger": "^9.2.0",
-    "@fastify/swagger-ui": "^5.1.0",
-    "fastify": "^5.1.0"
+    "@fastify/basic-auth": "^6.2.0",
+    "@fastify/swagger": "^9.5.1",
+    "@fastify/swagger-ui": "^5.2.3",
+    "fastify": "^5.5.0",
+    "openapi-types": "^12.1.3"
   },
   "types": "dist/swaig/src/index.d.ts",
   "devDependencies": {
