+++ packages/webrtc/package.json
@@ -42,7 +42,8 @@
     "test:integration:manual": "node run-integration-tests.js",
     "start:turn-server": "node -e \"require('./test/turnServer.js').MockTurnServer.getInstance().start().then(() => console.log('TURN server started')).catch(console.error)\"",
     "setup:playwright": "npx playwright install",
-    "prepublishOnly": "npm run build"
+    "prepublishOnly": "npm run build",
+    "build:integration": "node scripts/build-integration-bundle.js"
   },
   "dependencies": {
     "@signalwire/core": "4.3.1",
@@ -51,7 +52,8 @@
   "devDependencies": {
     "@playwright/test": "^1.54.2",
     "@types/node": "^20.19.10",
-    "node-turn": "^0.0.6"
+    "node-turn": "^0.0.6",
+    "esbuild": "^0.20.0"
   },
   "types": "dist/cjs/webrtc/src/index.d.ts"
 }
