+++ packages/webrtc/tsconfig.build.json
@@ -5,5 +5,5 @@
     "outDir": "dist/mjs"
   },
   "include": ["./src/**/*.ts", "../core/src/**/*.ts", "../webrtc/src/**/*.ts"],
-  "exclude": ["**/*.test.ts"]
+  "exclude": ["**/*.test.ts", "**/*.integration.test.ts", "**/*.spec.ts"]
 }
